Don’t create internal identity on account create
This removes the redudant creation of an `identity-internal` model when an `identity-account` was created. In addition, the identity and cookie management processes now allow for setting an `identity-account` directly, albeit without session data. This in no-way affects security or authentication, as an account still needs to be verified and credentials must be provided to move the account into a logged-in state.